requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
psycopg2==2.9.10
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
